requirements.txt
-c constraints.txt
lxml
click
bumpversion
flake8
tox
pytest
pytest-cov
pytest-mock
pytest-helpers-namespace
ipython
pytest-rpc
rpc-qtest-swagger-client
pytest-rpc
twine
requests
jinja2
future
gitpython
astor